Backed out changeset f8dbb1d2d07c (bug 1464400) for mochitest failures on test_font_whitelist.html. CLOSED TREE
UltraBlame original commit: 538a689e3487689416f0c06630a19cbc4ab193f7

+++ gfx/thebes/gfxFontFamilyList.h
@@ -23,7 +23,7 @@ namespace mozilla {
 
  
 
-enum FontFamilyType : uint8_t {
+enum FontFamilyType : uint32_t {
   eFamily_none = 0,
